chore: remove any casts and update agent guide
