+++ src/app/api/proposals/[id]/scry/route.ts
@@ -5,7 +5,8 @@ import { config } from '@/infrastructure/config'
 import { generateText } from 'ai'
 import { createOpenAI } from '@ai-sdk/openai'
 import { z } from 'zod'
-import { accumulateEffects, generateSkillTree } from '@/components/game/skills/procgen'
+import { generateSkillTree } from '@/components/game/skills/generate'
+import { accumulateEffects } from '@/components/game/skills/progression'
 import { rateLimit } from '@/middleware/rateLimit'
 import { buildGameContext } from '@/lib/gameContext'
