+++ src/app/play/PlayPageInternal.tsx
@@ -20,7 +20,9 @@ import { EdictsPanel, EdictSetting } from '@/components/game/EdictsPanel';
 import DistrictSprites, { District } from '@/components/game/districts';
 import { LeylineSystem, Leyline } from '@/components/game/LeylineSystem';
 import { createSupabaseBrowserClient } from '@/lib/supabase/browser';
-import { accumulateEffects, generateSkillTree, SkillNode } from '@/components/game/skills/procgen';
+import { generateSkillTree } from '@/components/game/skills/generate';
+import { accumulateEffects } from '@/components/game/skills/progression';
+import type { SkillNode } from '@/components/game/skills/types';
 import EffectsLayer from '@/components/game/EffectsLayer';
 import PreviewLayer from '@/components/game/PreviewLayer';
 import AmbientLayer from '@/components/game/AmbientLayer';
