+++ src/lib/__tests__/gameContext.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import { buildGameContext } from '../gameContext'
-import { generateSkillTree, accumulateEffects } from '../../components/game/skills/procgen'
+import { generateSkillTree } from '../../components/game/skills/generate'
+import { accumulateEffects } from '../../components/game/skills/progression'
 
 describe('buildGameContext', () => {
   it('summarizes buildings, routes, terrain and skills', () => {
